+++ src/Header/Range.php
@@ -3,32 +3,69 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header;
 
-use Innmind\Http\Header as HeaderInterface;
-use Innmind\Immutable\Sequence;
+use Innmind\Http\{
+    Header as HeaderInterface,
+    Exception\DomainException,
+};
+use Innmind\Immutable\{
+    Sequence,
+    Str,
+    Maybe,
+};
 
 /**
  * @psalm-immutable
  */
 final class Range implements HeaderInterface
 {
-    public function __construct(
-        private RangeValue $range,
+    /**
+     * @param int<0, max> $firstPosition
+     * @param int<0, max> $lastPosition
+     */
+    private function __construct(
+        private string $unit,
+        private int $firstPosition,
+        private int $lastPosition,
     ) {
     }
 
     /**
      * @psalm-pure
+     *
+     * @throws DomainException
      */
     public static function of(
         string $unit,
         int $firstPosition,
         int $lastPosition,
     ): self {
-        return new self(new RangeValue(
-            $unit,
-            $firstPosition,
-            $lastPosition,
-        ));
+        return self::maybe($unit, $firstPosition, $lastPosition)->match(
+            static fn($self) => $self,
+            static fn() => throw new DomainException($unit),
+        );
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @return Maybe<self>
+     */
+    public static function maybe(
+        string $unit,
+        int $firstPosition,
+        int $lastPosition,
+    ): Maybe {
+        if (
+            !Str::of($unit)->matches('~^\w+$~') ||
+            $firstPosition < 0 ||
+            $lastPosition < 0 ||
+            $firstPosition > $lastPosition
+        ) {
+            /** @var Maybe<self> */
+            return Maybe::nothing();
+        }
+
+        return Maybe::just(new self($unit, $firstPosition, $lastPosition));
     }
 
     #[\Override]
@@ -43,9 +80,25 @@ final class Range implements HeaderInterface
         return $this->header()->values();
     }
 
-    public function range(): RangeValue
+    public function unit(): string
     {
-        return $this->range;
+        return $this->unit;
+    }
+
+    /**
+     * @return int<0, max>
+     */
+    public function firstPosition(): int
+    {
+        return $this->firstPosition;
+    }
+
+    /**
+     * @return int<0, max>
+     */
+    public function lastPosition(): int
+    {
+        return $this->lastPosition;
     }
 
     #[\Override]
@@ -56,6 +109,14 @@ final class Range implements HeaderInterface
 
     private function header(): Header
     {
-        return new Header('Range', $this->range);
+        return new Header(
+            'Range',
+            new Value\Value(\sprintf(
+                '%s=%s-%s',
+                $this->unit,
+                $this->firstPosition,
+                $this->lastPosition,
+            )),
+        );
     }
 }