+++ tests/Header/RangeTest.php
@@ -6,29 +6,22 @@ namespace Tests\Innmind\Http\Header;
 use Innmind\Http\{
     Header\Range,
     Header,
-    Header\RangeValue
 };
 use Innmind\Immutable\Sequence;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\Attributes\DataProvider;
 
 class RangeTest extends TestCase
 {
     public function testInterface()
     {
-        $h = new Range(
-            $cr = new RangeValue('bytes', 0, 42),
-        );
+        $h = Range::of('bytes', 0, 42);
 
         $this->assertInstanceOf(Header::class, $h);
         $this->assertSame('Range', $h->name());
         $v = $h->values();
         $this->assertInstanceOf(Sequence::class, $v);
-        $this->assertSame($cr, $v->find(static fn() => true)->match(
-            static fn($first) => $first,
-            static fn() => null,
-        ));
         $this->assertSame('Range: bytes=0-42', $h->toString());
-        $this->assertSame($cr, $h->range());
     }
 
     public function testOf()
@@ -37,5 +30,32 @@ class RangeTest extends TestCase
 
         $this->assertInstanceOf(Range::class, $header);
         $this->assertSame('Range: bytes=0-42', $header->toString());
+    }
+
+    public function testValid()
+    {
+        $this->assertSame(
+            'Range: resources=0-42',
+            Range::of('resources', 0, 42)->toString(),
+        );
+    }
+
+    #[DataProvider('invalids')]
+    public function testReturnNothingWhenInvalidRangeValue($unit, $first, $last)
+    {
+        $this->assertNull(Range::maybe($unit, $first, $last)->match(
+            static fn($header) => $header,
+            static fn() => null,
+        ));
+    }
+
+    public static function invalids()
+    {
+        return [
+            ['', 0, 42],
+            ['foo', -1, 42],
+            ['foo', 0, -42],
+            ['foo', 100, 42],
+        ];
     }
 }